Added day 19 and day 20 code. Day 20 took over 24 hours in the end with breaks for sleep and life.
First solution was with p19w.pl which took ~50s to run.

p19.pl is an optimized version with more early outs, especially in the check for distance vector
matches - if the first value doesn't match any of the three then there's no point doing the
ordered comparison.

Day 20 was frustrating until I realized the infinity basically oscillates between on and off
and as long as you have a border and extend the working image using the right '#' or '.' for
each turn, you end up with a nice clean image turn after turn.